Fix peer dep. warnings
