+++ packages/design/src/utils/scrollbarSize.ts
@@ -1,0 +1,36 @@
+/*
+Copyright 2019-2020 Gravitational, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+// to cache scrollbar size value
+let size: number;
+
+export default function scrollbarSize(recalc?: boolean) {
+  if ((!size && size !== 0) || recalc) {
+    const scrollDiv = document.createElement('div');
+
+    scrollDiv.style.position = 'absolute';
+    scrollDiv.style.top = '-9999px';
+    scrollDiv.style.width = '50px';
+    scrollDiv.style.height = '50px';
+    scrollDiv.style.overflow = 'scroll';
+
+    document.body.appendChild(scrollDiv);
+    size = scrollDiv.offsetWidth - scrollDiv.clientWidth;
+    document.body.removeChild(scrollDiv);
+  }
+
+  return size;
+}
